Modified Component Creator for creation callbacks.
In the old system, a component registered itself with
the broadphases that should care about it.  This was alright
in a limited context, however became unruly under many circumstances.
This functionality is still provided for odd cases, but a much more
powerful system is in place.  Broadphases can now register callbacks
with the component creator, along with a type.  Whenever a component
is created which sublcasses this, the callback is called.  For example,
a Graphics Broadphase will register a callback for the "IRenderable" type.
Whenever a component is created which implements this interface, be it
StaticRenderable or AnimatedRenderable, the graphics broadphase will
automatically have a reference to the component.

Signed-off-by: Damian Lanningham <QRebound@gmail.com>